// UaRestGateway/uarestgateway.client/src/opcua/models/BrowseDirection.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OPC UA REST API
 * This API provides simple REST based access to an OPC UA server.
 *
 * The version of the OpenAPI document: 1.05.4
 * Contact: office@opcfoundation.org
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const BrowseDirection = {
    Forward: 0,
    Inverse: 1,
    Both: 2,
    Invalid: 3
} as const;
export type BrowseDirection = typeof BrowseDirection[keyof typeof BrowseDirection];


export function BrowseDirectionFromJSON(json: any): BrowseDirection {
    return BrowseDirectionFromJSONTyped(json, false);
}

export function BrowseDirectionFromJSONTyped(json: any, ignoreDiscriminator: boolean): BrowseDirection {
    return json as BrowseDirection;
}

export function BrowseDirectionToJSON(value?: BrowseDirection | null): any {
    return value as any;
}